style(cluster): move task releated into builder.go
