+++ pkg/cluster/builder.go
@@ -1,0 +1,454 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cluster
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"github.com/pingcap/tiup/pkg/cliutil"
+	"github.com/pingcap/tiup/pkg/cluster/executor"
+	operator "github.com/pingcap/tiup/pkg/cluster/operation"
+	"github.com/pingcap/tiup/pkg/cluster/spec"
+	"github.com/pingcap/tiup/pkg/cluster/task"
+	"github.com/pingcap/tiup/pkg/crypto"
+	"github.com/pingcap/tiup/pkg/environment"
+	"github.com/pingcap/tiup/pkg/meta"
+	pkgver "github.com/pingcap/tiup/pkg/repository/version"
+	"github.com/pingcap/tiup/pkg/set"
+)
+
+// Dynamic reload Prometheus configuration
+func buildDynReloadPromTasks(topo spec.Topology) []task.Task {
+	monitor := spec.FindComponent(topo, spec.ComponentPrometheus)
+	if monitor == nil {
+		return nil
+	}
+	instances := monitor.Instances()
+	if len(instances) == 0 {
+		return nil
+	}
+	var dynReloadTasks []task.Task
+	for _, inst := range monitor.Instances() {
+		dynReloadTasks = append(dynReloadTasks, task.NewBuilder().SystemCtl(inst.GetHost(), inst.ServiceName(), "reload", true).Build())
+	}
+	return dynReloadTasks
+}
+
+func buildScaleOutTask(
+	m *Manager,
+	clusterName string,
+	metadata spec.Metadata,
+	mergedTopo spec.Topology,
+	opt ScaleOutOptions,
+	sshConnProps *cliutil.SSHConnectionProps,
+	newPart spec.Topology,
+	patchedComponents set.StringSet,
+	optTimeout uint64,
+	sshTimeout uint64,
+	sshType executor.SSHType,
+	afterDeploy func(b *task.Builder, newPart spec.Topology),
+	final func(b *task.Builder, name string, meta spec.Metadata),
+) (task.Task, error) {
+	var (
+		envInitTasks       []task.Task // tasks which are used to initialize environment
+		downloadCompTasks  []task.Task // tasks which are used to download components
+		deployCompTasks    []task.Task // tasks which are used to copy components to remote host
+		refreshConfigTasks []task.Task // tasks which are used to refresh configuration
+	)
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+	specManager := m.specManager
+
+	tlsCfg, err := topo.TLSConfig(m.specManager.Path(clusterName, spec.TLSCertKeyDir))
+	if err != nil {
+		return nil, err
+	}
+
+	// Initialize the environments
+	initializedHosts := set.NewStringSet()
+	metadata.GetTopology().IterInstance(func(instance spec.Instance) {
+		initializedHosts.Insert(instance.GetHost())
+	})
+	// uninitializedHosts are hosts which haven't been initialized yet
+	uninitializedHosts := make(map[string]hostInfo) // host -> ssh-port, os, arch
+	newPart.IterInstance(func(instance spec.Instance) {
+		if host := instance.GetHost(); !initializedHosts.Exist(host) {
+			if _, found := uninitializedHosts[host]; found {
+				return
+			}
+
+			uninitializedHosts[host] = hostInfo{
+				ssh:  instance.GetSSHPort(),
+				os:   instance.OS(),
+				arch: instance.Arch(),
+			}
+
+			var dirs []string
+			globalOptions := metadata.GetTopology().BaseTopo().GlobalOptions
+			for _, dir := range []string{globalOptions.DeployDir, globalOptions.DataDir, globalOptions.LogDir} {
+				for _, dirname := range strings.Split(dir, ",") {
+					if dirname == "" {
+						continue
+					}
+					dirs = append(dirs, spec.Abs(globalOptions.User, dirname))
+				}
+			}
+			t := task.NewBuilder().
+				RootSSH(
+					instance.GetHost(),
+					instance.GetSSHPort(),
+					opt.User,
+					sshConnProps.Password,
+					sshConnProps.IdentityFile,
+					sshConnProps.IdentityFilePassphrase,
+					sshTimeout,
+					sshType,
+					globalOptions.SSHType,
+				).
+				EnvInit(instance.GetHost(), base.User, base.Group, opt.SkipCreateUser || globalOptions.User == opt.User).
+				Mkdir(globalOptions.User, instance.GetHost(), dirs...).
+				Build()
+			envInitTasks = append(envInitTasks, t)
+		}
+	})
+
+	// Download missing component
+	downloadCompTasks = convertStepDisplaysToTasks(BuildDownloadCompTasks(base.Version, newPart, m.bindVersion))
+
+	var iterErr error
+	// Deploy the new topology and refresh the configuration
+	newPart.IterInstance(func(inst spec.Instance) {
+		version := m.bindVersion(inst.ComponentName(), base.Version)
+		deployDir := spec.Abs(base.User, inst.DeployDir())
+		// data dir would be empty for components which don't need it
+		dataDirs := spec.MultiDirAbs(base.User, inst.DataDir())
+		// log dir will always be with values, but might not used by the component
+		logDir := spec.Abs(base.User, inst.LogDir())
+
+		deployDirs := []string{
+			deployDir, logDir,
+			filepath.Join(deployDir, "bin"),
+			filepath.Join(deployDir, "conf"),
+			filepath.Join(deployDir, "scripts"),
+		}
+		if topo.BaseTopo().GlobalOptions.TLSEnabled {
+			deployDirs = append(deployDirs, filepath.Join(deployDir, "tls"))
+		}
+		// Deploy component
+		tb := task.NewBuilder().
+			UserSSH(inst.GetHost(), inst.GetSSHPort(), base.User, sshTimeout, sshType, topo.BaseTopo().GlobalOptions.SSHType).
+			Mkdir(base.User, inst.GetHost(), deployDirs...).
+			Mkdir(base.User, inst.GetHost(), dataDirs...)
+
+		srcPath := ""
+		if patchedComponents.Exist(inst.ComponentName()) {
+			srcPath = specManager.Path(clusterName, spec.PatchDirName, inst.ComponentName()+".tar.gz")
+		}
+
+		if deployerInstance, ok := inst.(DeployerInstance); ok {
+			deployerInstance.Deploy(tb, srcPath, deployDir, version, clusterName, version)
+		} else {
+			// copy dependency component if needed
+			switch inst.ComponentName() {
+			case spec.ComponentTiSpark:
+				env := environment.GlobalEnv()
+				var sparkVer pkgver.Version
+				if sparkVer, _, iterErr = env.V1Repository().LatestStableVersion(spec.ComponentSpark, false); iterErr != nil {
+					return
+				}
+				tb = tb.DeploySpark(inst, sparkVer.String(), srcPath, deployDir)
+			default:
+				tb.CopyComponent(
+					inst.ComponentName(),
+					inst.OS(),
+					inst.Arch(),
+					version,
+					srcPath,
+					inst.GetHost(),
+					deployDir,
+				)
+			}
+		}
+		// generate and transfer tls cert for instance
+		if topo.BaseTopo().GlobalOptions.TLSEnabled {
+			ca, err := crypto.ReadCA(
+				clusterName,
+				m.specManager.Path(clusterName, spec.TLSCertKeyDir, spec.TLSCACert),
+				m.specManager.Path(clusterName, spec.TLSCertKeyDir, spec.TLSCAKey),
+			)
+			if err != nil {
+				iterErr = err
+				return
+			}
+			tb = tb.TLSCert(inst, ca, meta.DirPaths{
+				Deploy: deployDir,
+				Cache:  m.specManager.Path(clusterName, spec.TempConfigPath),
+			})
+		}
+
+		t := tb.ScaleConfig(clusterName,
+			base.Version,
+			m.specManager,
+			topo,
+			inst,
+			base.User,
+			meta.DirPaths{
+				Deploy: deployDir,
+				Data:   dataDirs,
+				Log:    logDir,
+			},
+		).Build()
+		deployCompTasks = append(deployCompTasks, t)
+	})
+	if iterErr != nil {
+		return nil, iterErr
+	}
+
+	hasImported := false
+
+	mergedTopo.IterInstance(func(inst spec.Instance) {
+		deployDir := spec.Abs(base.User, inst.DeployDir())
+		// data dir would be empty for components which don't need it
+		dataDirs := spec.MultiDirAbs(base.User, inst.DataDir())
+		// log dir will always be with values, but might not used by the component
+		logDir := spec.Abs(base.User, inst.LogDir())
+
+		// Download and copy the latest component to remote if the cluster is imported from Ansible
+		tb := task.NewBuilder()
+		if inst.IsImported() {
+			switch compName := inst.ComponentName(); compName {
+			case spec.ComponentGrafana, spec.ComponentPrometheus, spec.ComponentAlertmanager:
+				version := m.bindVersion(compName, base.Version)
+				tb.Download(compName, inst.OS(), inst.Arch(), version).
+					CopyComponent(compName, inst.OS(), inst.Arch(), version, "", inst.GetHost(), deployDir)
+			}
+			hasImported = true
+		}
+
+		// Refresh all configuration
+		t := tb.InitConfig(clusterName,
+			base.Version,
+			m.specManager,
+			inst,
+			base.User,
+			true, // always ignore config check result in scale out
+			meta.DirPaths{
+				Deploy: deployDir,
+				Data:   dataDirs,
+				Log:    logDir,
+				Cache:  specManager.Path(clusterName, spec.TempConfigPath),
+			},
+		).Build()
+		refreshConfigTasks = append(refreshConfigTasks, t)
+	})
+
+	// handle dir scheme changes
+	if hasImported {
+		if err := spec.HandleImportPathMigration(clusterName); err != nil {
+			return task.NewBuilder().Build(), err
+		}
+	}
+
+	// Deploy monitor relevant components to remote
+	dlTasks, dpTasks := buildMonitoredDeployTask(
+		m.bindVersion,
+		specManager,
+		clusterName,
+		uninitializedHosts,
+		topo.BaseTopo().GlobalOptions,
+		topo.BaseTopo().MonitoredOptions,
+		base.Version,
+		sshTimeout,
+		sshType,
+	)
+	downloadCompTasks = append(downloadCompTasks, convertStepDisplaysToTasks(dlTasks)...)
+	deployCompTasks = append(deployCompTasks, convertStepDisplaysToTasks(dpTasks)...)
+
+	builder := task.NewBuilder().
+		SSHKeySet(
+			specManager.Path(clusterName, "ssh", "id_rsa"),
+			specManager.Path(clusterName, "ssh", "id_rsa.pub")).
+		Parallel(false, downloadCompTasks...).
+		Parallel(false, envInitTasks...).
+		ClusterSSH(topo, base.User, sshTimeout, sshType, topo.BaseTopo().GlobalOptions.SSHType).
+		Parallel(false, deployCompTasks...)
+
+	if afterDeploy != nil {
+		afterDeploy(builder, newPart)
+	}
+
+	builder.
+		ClusterSSH(newPart, base.User, sshTimeout, sshType, topo.BaseTopo().GlobalOptions.SSHType).
+		Func("Save meta", func(_ *task.Context) error {
+			metadata.SetTopology(mergedTopo)
+			return m.specManager.SaveMeta(clusterName, metadata)
+		}).
+		Func("StartCluster", func(ctx *task.Context) error {
+			return operator.Start(ctx, newPart, operator.Options{OptTimeout: optTimeout}, tlsCfg)
+		}).
+		Parallel(false, refreshConfigTasks...).
+		Parallel(false, buildDynReloadPromTasks(metadata.GetTopology())...)
+
+	if final != nil {
+		final(builder, clusterName, metadata)
+	}
+
+	return builder.Build(), nil
+}
+
+type hostInfo struct {
+	ssh  int    // ssh port of host
+	os   string // operating system
+	arch string // cpu architecture
+	// vendor string
+}
+
+// Deprecated
+func convertStepDisplaysToTasks(t []*task.StepDisplay) []task.Task {
+	tasks := make([]task.Task, 0, len(t))
+	for _, sd := range t {
+		tasks = append(tasks, sd)
+	}
+	return tasks
+}
+
+func buildMonitoredDeployTask(
+	bindVersion spec.BindVersion,
+	specManager *spec.SpecManager,
+	clusterName string,
+	uniqueHosts map[string]hostInfo, // host -> ssh-port, os, arch
+	globalOptions *spec.GlobalOptions,
+	monitoredOptions *spec.MonitoredOptions,
+	version string,
+	sshTimeout uint64,
+	sshType executor.SSHType,
+) (downloadCompTasks []*task.StepDisplay, deployCompTasks []*task.StepDisplay) {
+	if monitoredOptions == nil {
+		return
+	}
+
+	uniqueCompOSArch := make(map[string]struct{}) // comp-os-arch -> {}
+	// monitoring agents
+	for _, comp := range []string{spec.ComponentNodeExporter, spec.ComponentBlackboxExporter} {
+		version := bindVersion(comp, version)
+
+		for host, info := range uniqueHosts {
+			// populate unique os/arch set
+			key := fmt.Sprintf("%s-%s-%s", comp, info.os, info.arch)
+			if _, found := uniqueCompOSArch[key]; !found {
+				uniqueCompOSArch[key] = struct{}{}
+				downloadCompTasks = append(downloadCompTasks, task.NewBuilder().
+					Download(comp, info.os, info.arch, version).
+					BuildAsStep(fmt.Sprintf("  - Download %s:%s (%s/%s)", comp, version, info.os, info.arch)))
+			}
+
+			deployDir := spec.Abs(globalOptions.User, monitoredOptions.DeployDir)
+			// data dir would be empty for components which don't need it
+			dataDir := monitoredOptions.DataDir
+			// the default data_dir is relative to deploy_dir
+			if dataDir != "" && !strings.HasPrefix(dataDir, "/") {
+				dataDir = filepath.Join(deployDir, dataDir)
+			}
+			// log dir will always be with values, but might not used by the component
+			logDir := spec.Abs(globalOptions.User, monitoredOptions.LogDir)
+			// Deploy component
+			t := task.NewBuilder().
+				UserSSH(host, info.ssh, globalOptions.User, sshTimeout, sshType, globalOptions.SSHType).
+				Mkdir(globalOptions.User, host,
+					deployDir, dataDir, logDir,
+					filepath.Join(deployDir, "bin"),
+					filepath.Join(deployDir, "conf"),
+					filepath.Join(deployDir, "scripts")).
+				CopyComponent(
+					comp,
+					info.os,
+					info.arch,
+					version,
+					"",
+					host,
+					deployDir,
+				).
+				MonitoredConfig(
+					clusterName,
+					comp,
+					host,
+					globalOptions.ResourceControl,
+					monitoredOptions,
+					globalOptions.User,
+					meta.DirPaths{
+						Deploy: deployDir,
+						Data:   []string{dataDir},
+						Log:    logDir,
+						Cache:  specManager.Path(clusterName, spec.TempConfigPath),
+					},
+				).
+				BuildAsStep(fmt.Sprintf("  - Copy %s -> %s", comp, host))
+			deployCompTasks = append(deployCompTasks, t)
+		}
+	}
+	return
+}
+
+func refreshMonitoredConfigTask(
+	specManager *spec.SpecManager,
+	clusterName string,
+	uniqueHosts map[string]hostInfo, // host -> ssh-port, os, arch
+	globalOptions spec.GlobalOptions,
+	monitoredOptions *spec.MonitoredOptions,
+	sshTimeout uint64,
+	sshType executor.SSHType,
+) []*task.StepDisplay {
+	if monitoredOptions == nil {
+		return nil
+	}
+
+	tasks := []*task.StepDisplay{}
+	// monitoring agents
+	for _, comp := range []string{spec.ComponentNodeExporter, spec.ComponentBlackboxExporter} {
+		for host, info := range uniqueHosts {
+			deployDir := spec.Abs(globalOptions.User, monitoredOptions.DeployDir)
+			// data dir would be empty for components which don't need it
+			dataDir := monitoredOptions.DataDir
+			// the default data_dir is relative to deploy_dir
+			if dataDir != "" && !strings.HasPrefix(dataDir, "/") {
+				dataDir = filepath.Join(deployDir, dataDir)
+			}
+			// log dir will always be with values, but might not used by the component
+			logDir := spec.Abs(globalOptions.User, monitoredOptions.LogDir)
+			// Generate configs
+			t := task.NewBuilder().
+				UserSSH(host, info.ssh, globalOptions.User, sshTimeout, sshType, globalOptions.SSHType).
+				MonitoredConfig(
+					clusterName,
+					comp,
+					host,
+					globalOptions.ResourceControl,
+					monitoredOptions,
+					globalOptions.User,
+					meta.DirPaths{
+						Deploy: deployDir,
+						Data:   []string{dataDir},
+						Log:    logDir,
+						Cache:  specManager.Path(clusterName, spec.TempConfigPath),
+					},
+				).
+				BuildAsStep(fmt.Sprintf("  - Refresh config %s -> %s", comp, host))
+			tasks = append(tasks, t)
+		}
+	}
+	return tasks
+}
